organize controller db files
